+++ src/Wfa2ClusterWorker.h
@@ -2,6 +2,7 @@
 #define OPTIMOTU_WFA2CLUSTERWORKER_H
 
 #include "DistClusterWorker.h"
+#include "SequenceView.h"
 
 class Wfa2ClusterWorker : public DistClusterWorker {
 protected:
@@ -10,7 +11,7 @@ protected:
     gap_open2 = 0, gap_extend2 = 1;
 public :
   Wfa2ClusterWorker(
-    const std::vector<std::string> &seq,
+    const SequenceSet &seq,
     ClusterAlgorithm &clust_algo,
     DivisiblePairGenerator::Builder & pgb,
     const int match = 0, const int mismatch = 1,
@@ -38,7 +39,7 @@ class Wfa2SplitClusterWorker : public Wfa2ClusterWorker {
   using Wfa2ClusterWorker::gap_extend2;
 public:
   Wfa2SplitClusterWorker(
-    const std::vector<std::string> &seq,
+    const SequenceSet &seq,
     ClusterAlgorithm &clust_algo,
     DivisiblePairGenerator::Builder & pgb,
     const int match = 0, const int mismatch = 1,
@@ -66,7 +67,7 @@ class Wfa2ConcurrentClusterWorker : public Wfa2ClusterWorker {
   using Wfa2ClusterWorker::gap_extend2;
   public:
   Wfa2ConcurrentClusterWorker(
-    const std::vector<std::string> &seq,
+    const SequenceSet &seq,
     ClusterAlgorithm &clust_algo,
     DivisiblePairGenerator::Builder & pgb,
     const int match = 0, const int mismatch = 1,